# ProcessGroupBaby: use pipe for improved performance